Update benchmark result of Medusa for batch size = 6 setup (#59)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->

Update benchmark result of Medusa for batch size = 6 setup
<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->
Update benchmark result of Medusa for batch size = 6 setup, and link the
plot for throughput and memory

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->
Done
<!-- 
Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- [x ] run `make test` to ensure correctness
- [x ] run `make checkstyle` to ensure code style
- [x ] run `make test-convergence` to ensure convergence

Co-authored-by: jaszhu <jzhu@linkedin.biz>