updated sl to be non-constexpr (#149)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
updated seqlen for rope to be non-constexpr

<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->
sl from constexpr to non-constexpr

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: RTX 3090
- [x] run `make test` to ensure correctness
- [x] run `make checkstyle` to ensure code style
- [x] run `make test-convergence` to ensure convergence