Optimize fused_linear_cross_entropy when weight does not require grads
Add some easy checks for `weight.requires_grad` to skip calculating weight gradients for this case. The weight gradient matrix can be pretty large, so this can also be a significant memory savings.

Also, skip the `.item()` call on `total_n_non_ignore` (the subsequent calculations work fine with the tensor form) to defer CUDA synchronization (otherwise it will wait for all the `torch.zeros` initializations on the preceding lines to synchronize, which may take a non-trivial amount of time.)