Cancel in-progress but out-of-date GPU actions (#289)
## Summary
When updating a PR branch, a lot of concurrent CICD actions can be
triggered. This creates a backlog and it takes time for the tests to
actually get around to running on the most recent commit. Github makes
it easy to cancel in-progress actions when new actions are trigged
within a group. This will help avoid wasting GPU time running CICD on
out of date commits.

Avoids cancelations on main, so that a complete CICD history is
available there.

- Hardware Type: <BLANK>
- [x] run `make test` to ensure correctness
- [x] run `make checkstyle` to ensure code style
- [x] run `make test-convergence` to ensure convergence

---------

Co-authored-by: Steven Shimizu <shimizust@gmail.com>