FIX: tl.program_id() does indeed not have a cast method in triton2.3.1 (#274)
## Summary
https://github.com/linkedin/Liger-Kernel/pull/251 casted program_id to
int64 in SwiGLU/GeGLU, but it's not compatible for triton2.3.1.

<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [x] run `make test` to ensure correctness
- [x] run `make checkstyle` to ensure code style
- [x] run `make test-convergence` to ensure convergence

Co-authored-by: Shao Tang <tangshao28@gmail.com>