Fix llama forward patch (#339)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->

The present version of liger kernel use `kwargs` in model forward
function, while in transformers 4.46.0-4.46.1, they pass the
`num_items_in_batch` parameter when `loss_kwargs` was in the model's
forward function [1][2], thus, we change the `kwargs` to `loss_kwargs`
to align with the transformers' implementation [3].

[1]
https://github.com/huggingface/transformers/blob/v4.46.1/src/transformers/trainer.py#L593
[2]
https://github.com/huggingface/transformers/blob/v4.46.1/src/transformers/trainer.py#L3620-L3625
[3]
https://github.com/huggingface/transformers/blob/v4.46.1/src/transformers/models/llama/modeling_llama.py#L1137-L1151

<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [ ] run `make test` to ensure correctness
- [ ] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence