Adds the CPO Alignment Loss Function (#382)
## Summary
CPO is almost the same as DPO with the major difference being that the
Reference Model in CPO is assumed to be a Uniform distribution. This
assumption leads to the cancellation of all terms related to the
reference model.


$$CPOLoss = -\log(\sigma(\beta\log(\pi_\theta(y_c|x)) -
\beta\log(\pi_\theta(y_r|x))))$$

This corresponds to equation 3 in the
[paper](https://arxiv.org/pdf/2401.08417). Additionally CPO also assumes
a scaling factor alpha for the NLL loss on the preferred response. In
TRL this corresponds to the CPOTrainer using a `loss_type="sigmoid"`

We also refactor the test cases for chunked loss functions to include a
generic `HFAlignmentLoss` base class that takes care some of the
plumbing work to correctly generate batches of input, calculate the
NLLoss etc. All future test cases can inherit from this class and just
implement the `alignment_loss` function to compare implementation in the
TRL lib versus the custom impl.

## Testing Done
A100-80G-SXM

Benchmark Results:

![Screenshot 2024-11-14 at 5 17
42 PM](https://github.com/user-attachments/assets/64deda54-e48a-4c6c-a704-073f16a72085)
![Fused Linear CPO Loss
Speed](https://github.com/user-attachments/assets/d231b38c-63d7-440b-9a64-56c95e819c89)


- Hardware Type: <BLANK>
- [X] run `make test` to ensure correctness
- [X] run `make checkstyle` to ensure code style
- [X] run `make test-convergence` to ensure convergence

---------

Co-authored-by: shivam15s <shivam15800@gmail.com>