[AMD] [CI] Clean up `amd-ci` (#436)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
This is to clean up the `amd-ci.yml` setup steps.

<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [x] run `make test` to ensure correctness
- [x] run `make checkstyle` to ensure code style
- [x] run `make test-convergence` to ensure convergence

---------

Co-authored-by: tjtanaa <tunjian.tan@embeddedllm.com>