Fix Preference Loss and Refactor for Readability  (#484)
## Summary

Thanks to @winglian and @shivam15s noticed and fixed this
https://github.com/linkedin/Liger-Kernel/pull/481.

This PR suggests negating the preference loss terms to align with the
formulas in the docstrings, while maintaining the base preference
structure as `nll_loss + preference_loss`. This would make our loss
computations more consistent since both terms would represent losses to
be minimized.

[UPDATE: It seems like being addressed now in
[here](https://github.com/linkedin/Liger-Kernel/commit/3205342a6a7209c55ca3a4bd97e986961fdc792e#diff-3048cb37b97e27515852c200994f3257b8ae33a465421d05184713377c0895b1R150)]
This PR also tightened the tolerance in case of encountering a similar
issue.

<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [X] run `make test` to ensure correctness
- [X] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence

---------

Signed-off-by: Austin Liu <austin362667@gmail.com>
Co-authored-by: Wing Lian <wing@axolotl.ai>
Co-authored-by: Shivam Sahni <shivam15800@gmail.com>