Refactor Temperature Scaling in Distillation Loss (#444)
## Summary

Addressed the part of issue raised in
https://github.com/linkedin/Liger-Kernel/issues/441

Moving the scale temperature outside the `distillation_loss_fn` is fine
as well. Keep the `loss_fn` simpler, and the rest can be handled in the
`forward` function beforehand. Thanks to the advice by @Tcc0403

<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [ ] run `make test` to ensure correctness
- [X] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence

---------

Signed-off-by: Austin Liu <austin362667@gmail.com>