Add TVD Loss Kernel (#324)
## Summary
Resolves [#281](https://github.com/linkedin/Liger-Kernel/issues/281).
Implements the TVD (Total Variation Distance) kernel by computing both
the loss and gradient in the forward pass.

## Testing Done
Implemented tests to verify that the results of the forward and backward
passes match the Torch implementation. Additionally, added a script to
benchmark the memory usage and speed of the Liger implementation
compared to Torch, with the results shown below.


![tvd_speed](https://github.com/user-attachments/assets/05080030-81ae-4751-aba2-f001a5144072)


![tvd_memory](https://github.com/user-attachments/assets/d7dbb9d9-5fd5-4ffe-aa12-8ba785e09857)

- Hardware Type: Nvidia H100 (80GB PCIe)
- [x] run `make test` to ensure correctness
- [x] run `make checkstyle` to ensure code style
- [x] run `make test-convergence` to ensure convergence

---------

Co-authored-by: Shao Tang <tangshao28@gmail.com>