Fix Rope Test (#577)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
Closes https://github.com/linkedin/Liger-Kernel/issues/565


<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

If the utility functions are being used by both the benchmark and test
directories, it's a good idea to move the shared code to src/ so that
both benchmark/ and test/ can access it, but without mixing it directly
into the test/ directory.

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->
make test
<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [x] run `make test` to ensure correctness
- [x] run `make checkstyle` to ensure code style
- [x] run `make test-convergence` to ensure convergence