Compare means of aggregated outputs in KTO tests (#595)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
This change modifies the testing code to compare the mean of aggregated
outputs in kto tests. This prevents the problem of error accumulation
when we sum the metrics.

<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type:  H100-80GB-HBM3
- [x] run `make test` to ensure correctness
- [x] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence