Zhasun/grpo add ref logits (#635)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
Add the ref_log_probs as input. and if the log probs is not none, we
don't need to re-calculate it using ref_input.

Add it in corresponding unit test as well.
<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [X] run `make test` to ensure correctness
- [X] run `make checkstyle` to ensure code style
- [X] run `make test-convergence` to ensure convergence