add XPU tuning to JSD (#649)
## Summary
Tuning on XPU: In JSD, if device is XPU, set MAX_FUSED_SIZE to 4096
instead of default 65536. This gives slightly better performance on XPU.
This PR is similar to PR #647.

<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: Intel(R) Data Center GPU Max 1550
- [x] run `make test` to ensure correctness
- [x] run `make checkstyle` to ensure code style
- [x] run `make test-convergence` to ensure convergence

---------

Co-authored-by: Shivam Sahni <shivam15800@gmail.com>
Co-authored-by: Shao Tang <tangshao28@gmail.com>