Fix LigerCrossEntropy reduction='none' (#680)
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
Resolves #678.

Current CE and FLCE can't calculate the gradients correctly when
reduction='none'. In this scenario, `grad_output` is not just a scalar
and it should be multiplied to each row respectively isntead.
<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [ ] run `make test` to ensure correctness
- [ ] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence

---------

Signed-off-by: Tcc0403 <76503978+Tcc0403@users.noreply.github.com>
Co-authored-by: Shivam Sahni <shivam15800@gmail.com>