Use logits_to_keep logic for training runs (#696)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
This PR fixes #694. 

It adds a change to slice the outputs according to `logits_to_keep`
before calculating the loss during the training.

<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [ ] run `make test` to ensure correctness
- [ ] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence

Co-authored-by: Shivam Sahni <shivam15800@gmail.com>