[Fix] Deploy the Mkdocs site to github pages (#724)
## Summary
Follow-up of #722

I've added a separate folder 'doc-site' to compile the site and modified
the mkdocs.yml to that folder path so that when the github workflow is
executed, the site content will be fetched from 'doc-site'. It is
important to note that 'doc-site' cannot be placed under the docs folder
since it leads to an error.

I've also updated the Makefile to point towards the correct path as
well.

The deployed site is from my fork of the repo:
https://paragekbote.github.io/Liger-Kernel/

Would like to add me a README note about the site or update the About
section with the website link?

Could you please let me know?

cc: @shimizust

<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [ ] run `make test` to ensure correctness
- [ ] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence