Fix RMS norm patching (#741)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
Fixes #739.

<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->
Creates new classes for various RMS norms and removes the use of
`partial` for RMS norms.

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->
Fixes errors of the form:
```
TypeError: isinstance() arg 2 must be a type, a tuple of types, or a union
```
for glm4, olmo2, gemma1 and gemma2.

However, we are now seeing errors when matching the actual logits with
gemma models, which can be tracked separately in
https://github.com/linkedin/Liger-Kernel/issues/729.
<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [ ] run `make test` to ensure correctness
- [ ] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence