Adding a test to run one benchmark when anything is pushed to the main branch (#744)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
Trying to run just the KTO loss for now for testing purposes. It will
run after any thing is pushed to the main branch
<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [ ] run `make test` to ensure correctness
- [ ] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence