Adding the benchmark results to gh-pages branch (#747)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
This is pr for adding the data generated from the benchmark scripts to
the gh-pages branch. It adds in this fashion:
latest_main_branch_commit_hash/benchmark.csv. 

<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [ ] run `make test` to ensure correctness
- [ ] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence