UI for the benchmarking results (#758)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
This is the UI that compares two different commit's benchmarking
results.
The data is fetched from the gh-pages branch where data is store as per
commit.
<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [ ] run `make test` to ensure correctness
- [ ] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>