Automate benchmarking - changed the scripts for backward pass. (#759)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
The scripts are updated so that we get benchmarking results on backward
pass as well.
All benchmark scripts were run on a H100, it took around 55 minutes to
run.

There are total of 26 scripts, so each generates a plot for speed -
forward, backward and full pass
for memory - full pass. Total of 104.
And the output (csv file) was of 460kB.
<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [ ] run `make test` to ensure correctness
- [ ] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence

---------

Co-authored-by: Manan Shah <manashah@linkedin.com>
Co-authored-by: Vaibhav Jindal <vaibhav.jndl@gmail.com>