Trying out logprobs and top logprobs for testing rather than logits. (#745)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
Just testing out logprobs as mentioned in #742 
It worked for the models where the test using logits was not working. 
Also, tried to setup 1e-1 tolerance for qwen (previously 1) and it
passed.
<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [ ] run `make test` to ensure correctness
- [ ] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence