Liger support for Llama4 (#740)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
Adding liger support to llama4 (text and multimodal). Liger rope is not
supported yet for llama4 as it has a different implementation.
The tolerance are set higher than usual to pass the tests. 

<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: NVIDIA H100 80GB HBM3
- [ ] run `make test` to ensure correctness
- [x] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence

---------

Co-authored-by: Shao Tang <tangshao28@gmail.com>
Co-authored-by: Manan Shah <manashah@linkedin.com>
Co-authored-by: Vaibhav Jindal <vaibhav.jndl@gmail.com>