[Model] Liger support for SmolLM3 (#798)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
This PR adds Liger support for the SmolLM3 model.
<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->
The model architecture is similar to that of llama, so I was able to use
most of their ops.

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: A100-80G-PCIe
- [x] run `make test` to ensure correctness
- [x] run `make checkstyle` to ensure code style
- [x] run `make test-convergence` to ensure convergence

---------

Co-authored-by: Steven Shimizu <shimizust@gmail.com>