Skip smollm3 tests in tests-bwd (#821)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
Skip smollm3 tests in tests-bwd as it is only available for
transformers>=4.53.0. This will help fix tests-bwd on nvidia.
 
<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [x] run `make test` to ensure correctness
- [x] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence