CI tests fix (#847)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
This fixes the CI test, logging is removed from the huggingface
transformers which was causing errors.

https://github.com/huggingface/transformers/commit/019b74977d7469af3a86d3f5bbbb02ffe62af94d#diff-73dfbce7b723db29b1149d4de45a857b2b4ecab7387af1742e5c2686b2175fd1

<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [ ] run `make test` to ensure correctness
- [ ] run `make checkstyle` to ensure code style
- [x] run `make test-convergence` to ensure convergence