Fixing URL issues (#849)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
The github rawusercontent which is used to fetch the data for the
benchmark dashboard sends 404 error sometimes because of which the
benchmark dashboard does not show data. So putting in a fallback URL
from cdn.jsdeliver.net.
<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [ ] run `make test` to ensure correctness
- [ ] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Manan Shah <manashah@linkedin.com>