fix(phi3): update monkey patch for `Phi3ForCausalLM` (#837)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
Liger backward support for transformers has been updated to 4.49.0, so
we can remove most reprecated functions that were made for
`SUPPORTED_TRANSFORMER_VERSION=4.46,1` (we should probably bump it for
next release too).

This PR merges most patching logics into one that should be able to work
with both 4.49.0 and the latest version (4.54.0)
<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [ ] run `make test` to ensure correctness
- [ ] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence

---------

Signed-off-by: Tcc0403 <76503978+Tcc0403@users.noreply.github.com>
Co-authored-by: Shao Tang <tangshao28@gmail.com>
Co-authored-by: Steven Shimizu <shimizust@gmail.com>