Fix ci (#853)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
This fixes the CI.
The mllama model does not support the accum_dtype param for transfomers
version 4.49.0.
The qwen model required out_hidden_size param in the config.
<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: H100
- [x] run `make test` to ensure correctness
- [x] run `make checkstyle` to ensure code style
- [x] run `make test-convergence` to ensure convergence

---------

Co-authored-by: Manan Shah <manashah@linkedin.com>
Co-authored-by: Vaibhav Jindal <vaibhav.jndl@gmail.com>