Fix reduction param validation for flce (#879)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->

This is a fix for Issue #872. @shimizust wants me to submit this fix. 

<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->
@shimizust Could you have a look at the fix?

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [ ] run `make test` to ensure correctness
- [ ] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence